Iniciada correção da classe pessoa e classe users
